Add TorrentMover class and refactor import statements.
Added a new class, TorrentMover, to handle file type determination and path evaluation in order to improve code organization and separation of concerns.

Also, to enforce the python naming convention of CamelCase for classes, the filename of 'Torrent', 'Series', and 'Config' classes were capitalized. Other files importing these classes had to be updated to reflect these changes.

Additionally, added logic to stop tracker checks once the tracker is set in the Torrent class to enhance performance.